+++ intellijad/src/java/net/stevechaloner/intellijad/config/ProcyonConfig.java
@@ -1,0 +1,4 @@
+package net.stevechaloner.intellijad.config;
+
+public class ProcyonConfig {
+}
